change directory of main.py and clean file
